Removed redundant label renderer
